Optimized again, works at 46kHz on the stk :D

+++ include/Player.h
@@ -53,7 +53,7 @@ void MOD_Player_division(MOD_Player* player);
 
 MOD_Player_Channel* MOD_Player_Channel_create();
 void MOD_Player_Channel_free(MOD_Player_Channel* channel);
-int16_t MOD_Player_Channel_step(MOD_Player_Channel* player_channel, MOD_Player* player, MOD* mod);
+int32_t MOD_Player_Channel_step(MOD_Player_Channel* player_channel, MOD_Player* player, MOD* mod);
 
 void MOD_Player_Channel_set_volume(MOD_Player_Channel* player_channel, int volume);
 void MOD_Player_Channel_process_effect(MOD_Player_Channel* player_channel, MOD_Player* player, MOD* mod, int effect);
